pyproject.toml
[tool.black]
line-length = 88
exclude = '''
/(
  | migrations
  | .venv
)/
'''

[tool.flake8]
ignore = ["F405", "E501", "W503", "E731", "E741",]
max-line-length = 88
exclude = [
    ".git", "migrations", "node_modules", "docs", ".venv", "frontend"
]
statistics = false

[tool.isort]
line_length = 88
skip_glob = "*/migrations/.py"
multi_line_output = 3
profile = "black"
filter_files = true
include_trailing_comma = true
force_grid_wrap = 2

[tool.poetry]
name = "task-tile-app"
version = "0.1.0"
description = ""
authors = ["Piwero <piwero@gmail.com>"]

[tool.poetry.dependencies]
Django = "2.2.24"
djangorestframework = "^3.12.4"
python = "^3.9"
python-decouple = "^3.4"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.9"
isort = "^5.9.3"
flake9 = "3.8.3"
pre-commit = "^2.13.0"
black = "^21.7b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
